+++ changelog.md
@@ -18,6 +18,7 @@
  ### Units
    * Undead variations for Falcon, Giant Rat, serpents, and Gorer/Tusklet
  ### User interface
+   * Text labels now use the hyperlink mouse cursor while hovering links.
  ### WML Engine
  ### Miscellaneous and Bug Fixes
    * Fixed several possible crashes in wmllint